Data file containing possible income statement keys from Yahoo Finance API

A dataset containing the possible keys that can be found in income statement data retrieved from Yahoo Finance's API. These keys are used to identify and extract specific income statement line items.

Usage

valid_income_keys

Format

A character vector containing 103 possible income statement keys

Source

Yahoo Finance API

Examples

# View all possible income statement keys
data(valid_income_keys)
head(valid_income_keys)
#> [1] "Amortization"                            
#> [2] "AmortizationOfIntangiblesIncomeStatement"
#> [3] "AverageDilutionEarnings"                 
#> [4] "BasicAccountingChange"                   
#> [5] "BasicAverageShares"                      
#> [6] "BasicContinuousOperations"               

# Check if a specific key exists
"NetIncome" %in% valid_income_keys
#> [1] TRUE